% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_SUCRA_plot.R
\name{forest_SUCRA_plot}
\alias{forest_SUCRA_plot}
\title{Forest-SUCRA plot of net-meta model for MD, HR, and RR}
\usage{
forest_SUCRA_plot(
  bmt,
  x.lab,
  title.size = 15,
  font.size = 5,
  x1,
  x2,
  digits = 2,
  sucra,
  labels
)
}
\arguments{
\item{bmt}{bayesian net-meta result from model_gemtc}

\item{x.lab}{label in the x-axis}

\item{title.size}{title size}

\item{font.size}{font size}

\item{x1}{the left xlim}

\item{x2}{the right xlim}

\item{digits}{digits of the results}

\item{sucra}{SUCRA values}

\item{labels}{label of the whole plot}
}
\value{
ggplot object
}
\description{
To give the forest plot with SUCRA of bayesian results from model_gemtc
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

sucra<-SUCRA(bmt1)

p <- forest_SUCRA_plot(
bmt=bmt1,
x.lab="X caption",
title.size =15,
font.size =5,
x1=0,
x2=40,
digits = 2,
sucra=sucra,
labels="Title")

}
