% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_table_plot.R
\name{forest_table_plot}
\alias{forest_table_plot}
\title{Forest-table plot of net-meta model for MD, HR, and RR}
\usage{
forest_table_plot(
  bmt,
  nmt = NULL,
  x.lab,
  title.size = 15,
  font.size = 5,
  x1,
  x2,
  digits = 2,
  type = c("bayesian", "all")
)
}
\arguments{
\item{bmt}{bayesian net-meta result from model_gemtc}

\item{nmt}{frequentist net-meta result from model_nemeta}

\item{x.lab}{label in the x-axis}

\item{title.size}{title size}

\item{font.size}{font size}

\item{x1}{the left xlim}

\item{x2}{the right xlim}

\item{digits}{digits of the results}

\item{type}{the type of plots, bayesian-bayesian result; all-bayesian, freq, and prediction interval}
}
\value{
ggplot object
}
\description{
To give the forest plot with table of bayesian results from model_gemtc or
combined forest plot of bayesian results, frequentist results, and prediction interval from model_gemtc and model_netmeta.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

p<-forest_table_plot(
bmt=bmt1,
x.lab = "X",
x1=0,
x2=40,
title.size =15,
font.size =5,
digits = 2,
type="bayesian")

nmt1 <- model_netmeta(long.data = LDT1,
                      treatment=LDT1$treatment,
                      id.treatments = trt1,
                      reference = "A",
                      outcome = "HR")

# View(nmt1)

p<-forest_table_plot(
bmt=bmt1,
nmt=nmt1,
x.lab = "X",
x1=0,
x2=50,
title.size =15,
font.size =5,
digits = 2,
type="all")

}
