% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_plot.R
\name{funnel_plot}
\alias{funnel_plot}
\title{Funnel plot from frequentist net-meta}
\usage{
funnel_plot(nmt, font.size, text.size, text.x, cap, x1, x2)
}
\arguments{
\item{nmt}{results from model_netmeta function}

\item{font.size}{font size}

\item{text.size}{text size}

\item{text.x}{text position}

\item{cap}{caption}

\item{x1}{left xlim}

\item{x2}{right xlim}
}
\value{
a ggplot object
}
\description{
To plot the funnel plot from frequentist net-meta
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

trt1$label <- paste0(trt1$id,"-", trt1$description)
LDT1$label <- factor(LDT1$treatment, labels = trt1$label)

trt1$label <- paste0(trt1$id,"-", trt1$description)
LDT1$label <- factor(LDT1$treatment, labels = trt1$label)

nmt.lab1 <- model_netmeta(
  long.data=LDT1[,-2],
  treatment=LDT1$label,
  id.treatments=trt1,
  reference = trt1$label[1],
  outcome="HR")

p<-funnel_plot(
  nmt.lab1,
  font.size=10,
  text.size=4,
  text.x=0,
  x1=-1,
  x2=3,
  cap="HR")

}
