% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRADE_plot.R
\name{GRADE_plot}
\alias{GRADE_plot}
\title{GRADE plot}
\usage{
GRADE_plot(
  RESULT,
  font.size,
  lab.size,
  font.family = c("Helvetica", "Times New Roman")
)
}
\arguments{
\item{RESULT}{GRADE matrix result}

\item{font.size}{font size}

\item{lab.size}{text size}

\item{font.family}{"Helvetica" or "Times New Roman"}
}
\value{
ggplot object
}
\description{
To plot GRADE result.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

nmt1 <- model_netmeta(long.data = LDT1,
treatment=LDT1$treatment,
id.treatments = trt1,
reference = "A",
outcome = "HR")

name1 <- NULL
for(i in 1:(length(trt1$id)-1)){
  name1 <- c(name1, paste0(trt1$id[i], ":", trt1$id[-(1:i)]))
}

study.CM1 <- studyCM_matrix(name1, nmt1)

study.assess1 <- read.csv(system.file("extdata", "HR_SH_A.csv", package = "net.meta"))

RB.comp1 <- rep(0, nrow(study.CM1)) #1 Yes, 0 no

inconsist1<-gemtc::mtc.nodesplit(
bmt1$mtc.net,
likelihood = "poisson",
link = "log",
linearModel = "random",
dic=TRUE)

RESULT.B1 <- GRADE_table(
  study.CM1,
  bmt1,
  nmt1,
  rob=study.assess1$ROB,  ## define Rob per study
  ind=study.assess1$IND,  ## define Indirectness per study
  report.bias=RB.comp1,
  effect.size = "ratio",
  clinical.effect.size=1.25,
  inconsis = inconsist1,
  model="Bayes")

p<-GRADE_plot(RESULT.B1, font.size=4, lab.size=12)

}
