% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_netmeta.R
\name{model_netmeta}
\alias{model_netmeta}
\title{Frequentist net-meta model for MD, HR, and RR}
\usage{
model_netmeta(
  long.data,
  treatment,
  id.treatments,
  reference,
  outcome = c("MD", "HR", "RR")
)
}
\arguments{
\item{long.data}{data.frame to be analyzed should be formatted in long format}

\item{treatment}{treatment names or labels}

\item{id.treatments}{data.frame to specify the id and treatments}

\item{reference}{the referential id in the net-meta}

\item{outcome}{the outcome should be MD-mean difference, HR-hazard ratio, and RR-risk ratio}
}
\value{
Summary list of the results
}
\description{
This function is the warpper of mtc.network, mtc.model, mtc.run, and relative.effect from gemtc package. \cr
R packgae gemtc and rjags are required.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "mHR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "mHR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

trt1$label <- paste0(trt1$id,"-", trt1$description)
LDT1$label <- factor(LDT1$treatment, labels = trt1$label)


nmt.lab1 <- model_netmeta(
  long.data=LDT1[,-2],
  treatment=LDT1$label,
  id.treatments=trt1,
  reference = "A",
  outcome="HR")

# View(nmt.lab1)
}
