% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoB_plot.R
\name{RoB_plot}
\alias{RoB_plot}
\title{Risk of bias bar plot}
\usage{
RoB_plot(study.CM, rob.ind, lab.size)
}
\arguments{
\item{study.CM}{contribution matrix}

\item{rob.ind}{risk of bias or indirectness}

\item{lab.size}{label size}
}
\value{
ggplot object
}
\description{
To plot risk of bias result.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

nmt1 <- model_netmeta(long.data = LDT1,
treatment=LDT1$treatment,
id.treatments = trt1,
reference = "A",
outcome = "HR")

name1 <- NULL
for(i in 1:(length(trt1$id)-1)){
  name1 <- c(name1, paste0(trt1$id[i], ":", trt1$id[-(1:i)]))
}

study.CM1 <- studyCM_matrix(name1, nmt1)

study.assess1 <- read.csv(system.file("extdata", "HR_SH_A.csv", package = "net.meta"))

p<-RoB_plot(study.CM1, study.assess1$ROB, lab.size=10)

p<-RoB_plot(study.CM1, study.assess1$IND, lab.size=10)

}
