% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCM_matrix.R
\name{studyCM_matrix}
\alias{studyCM_matrix}
\title{Contribution matrix of frequentist net-meta}
\usage{
studyCM_matrix(comp.all, nmt)
}
\arguments{
\item{comp.all}{all comparisoin}

\item{nmt}{results from model_netmeta function}
}
\value{
matrix
}
\description{
To calculate the contribution matrix.
}
\examples{

LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

nmt1 <- model_netmeta(long.data = LDT1,
treatment=LDT1$treatment,
id.treatments = trt1,
reference = "A",
outcome = "HR")

name1 <- NULL
for(i in 1:(length(trt1$id)-1)){
  name1 <- c(name1, paste0(trt1$id[i], ":", trt1$id[-(1:i)]))
}

study.CM1 <- studyCM_matrix(name1, nmt1)

}
