% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUCRA.R
\name{SUCRA}
\alias{SUCRA}
\title{SUCRA from bayesian model}
\usage{
SUCRA(bmt, digits = 3)
}
\arguments{
\item{bmt}{result from model_gemtc}

\item{digits}{digits of the result}
}
\value{
SUCRA result
}
\description{
To calculate SUCRA from bayesian model.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

# View(bmt1)

SUCRA(bmt1)

}
