% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_table_plot.R
\name{forest_table_plot}
\alias{forest_table_plot}
\title{Forest-table plot of net-meta model for MD, HR, and RR}
\usage{
forest_table_plot(
  type = c("bayesian", "all"),
  bmt,
  nmt = NULL,
  digits = 2,
  x1,
  x2,
  x.lab,
  title.size,
  text.size,
  point.size = 4,
  labels = NULL,
  plot.scale = 0.9,
  font.family = "Helvetica"
)
}
\arguments{
\item{type}{the type of plots, bayesian-bayesian result; all-bayesian, freq, and prediction interval}

\item{bmt}{bayesian net-meta result from model_gemtc}

\item{nmt}{frequentist net-meta result from model_nemeta}

\item{digits}{digits of the results}

\item{x1}{the left xlim}

\item{x2}{the right xlim}

\item{x.lab}{label in the x-axis}

\item{title.size}{title size}

\item{text.size}{font size}

\item{point.size}{point size}

\item{labels}{label of the whole plot}

\item{plot.scale}{the scaled size of the whole plot}

\item{font.family}{"Helvetica" or "Times New Roman"}
}
\value{
ggplot object
}
\description{
To give the forest plot with table of bayesian results from model_gemtc or
combined forest plot of bayesian results, frequentist results, and prediction interval from model_gemtc and model_netmeta.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

forest_table_plot(
type="bayesian",
bmt=bmt1,
x.lab = "X caption",
x1=0,
x2=40,
title.size =12,
text.size =4,
digits = 2,
font.family="Helvetica")

nmt1 <- model_netmeta(long.data = LDT1,
                      treatment=LDT1$treatment,
                      id.treatments = trt1,
                      reference = "A",
                      outcome = "HR")

# View(nmt1)

forest_table_plot(
type="all",
bmt=bmt1,
nmt=nmt1,
x.lab = "X caption",
x1=0,
x2=50,
title.size =12,
text.size =4,
digits = 2)

}
