% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRADE_table.R
\name{GRADE_table}
\alias{GRADE_table}
\title{GRADE table}
\usage{
GRADE_table(
  study.CM,
  bmt,
  nmt,
  rob,
  ind,
  report.bias,
  effect.size = c("diff", "ratio"),
  clinical.effect.size,
  inconsis,
  model = c("Bayes", "Freq")
)
}
\arguments{
\item{study.CM}{contribution matrix from sutdyCM_matrix}

\item{bmt}{result from bayesian net-meta model_gemtc}

\item{nmt}{result from frequentist net-meta model_netmmeta}

\item{rob}{risk of bias}

\item{ind}{indirectness}

\item{report.bias}{reported bias}

\item{effect.size}{two types: diff-difference; ratio-ratio}

\item{clinical.effect.size}{clinical effect size}

\item{inconsis}{inconsistence from mtc.nodesplit}

\item{model}{Bayes-bayesian net-meta model, Freq-frequentist net-meta model}
}
\value{
GRADE table matrix
}
\description{
To calculate the GRADE table.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

nmt1 <- model_netmeta(long.data = LDT1,
treatment=LDT1$treatment,
id.treatments = trt1,
reference = "A",
outcome = "HR")

name1 <- NULL
for(i in 1:(length(trt1$id)-1)){
  name1 <- c(name1, paste0(trt1$id[i], ":", trt1$id[-(1:i)]))
}

study.CM1 <- studyCM_matrix(name1, nmt1)

study.assess1 <- read.csv(system.file("extdata", "HR_SH_A.csv", package = "net.meta"))

RB.comp1 <- rep(0, nrow(study.CM1)) #1 Yes, 0 no

inconsist1<-gemtc::mtc.nodesplit(
bmt1$mtc.net,
likelihood = "poisson",
link = "log",
linearModel = "random",
dic=TRUE)

RESULT.F1 <- GRADE_table(
  study.CM1,
  bmt=bmt1,
  nmt=nmt1,
  rob=study.assess1$ROB,
  ind=study.assess1$IND,
  report.bias=RB.comp1,
  effect.size = "ratio",
  clinical.effect.size=1.25,
  inconsis = inconsist1,
  model="Freq")

RESULT.B1 <- GRADE_table(
  study.CM1,
  bmt1,
  nmt1,
  rob=study.assess1$ROB,  ## define Rob per study
  ind=study.assess1$IND,  ## define Indirectness per study
  report.bias=RB.comp1,
  effect.size = "ratio",
  clinical.effect.size=1.25,
  inconsis = inconsist1,
  model="Bayes")

}
