% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_score_plot.R
\name{net_score_plot}
\alias{net_score_plot}
\title{Network and SUCRA plots}
\usage{
net_score_plot(
  nmt,
  sucra,
  scl = 1,
  ofs = 0.05,
  lab,
  font.size,
  pos = 0.5,
  labels = ""
)
}
\arguments{
\item{nmt}{results from model_netmeta function}

\item{sucra}{sucra results from SUCRA function}

\item{scl}{additional space added outside of edges in the net graph}

\item{ofs}{distance between edges in the net graph}

\item{lab}{an optional vector with treatment labels in the net graph}

\item{font.size}{font size}

\item{pos}{a [0, 1] specifying the position of the number of studies on the lines connecting treatments (edges)}

\item{labels}{label of the whole plots}
}
\value{
a ggplot object
}
\description{
To plot the network graph with SUCRA.
}
\examples{
LDT1 <- read.csv(system.file("extdata", "HR_SH_D.csv", package = "net.meta"))
trt1 <- read.table(system.file("extdata", "HR_SH_D.txt", package = "net.meta"),
    header=TRUE,quote = '"', stringsAsFactors=FALSE)

trt1$description <- factor(trt1$description, trt1$description)
LDT1$study <- factor(LDT1$study, unique(LDT1$study))

bmt1 <- model_gemtc(
long.data=LDT1,
id.treatments=trt1,
reference="A",
outcome="HR",
mtc.n.adapt = 5000, mtc.n.iter = 10000, mtc.thin = 20)

# View(bmt1)

sucra1 <-SUCRA(bmt1)

trt1$label <- paste0(trt1$id,"-", trt1$description)
LDT1$label <- factor(LDT1$treatment, labels = trt1$label)


nmt1 <- model_netmeta(long.data = LDT1,
                      treatment=LDT1$treatment,
                      id.treatments = trt1,
                      reference = "A",
                      outcome = "HR")

# View(nmt1)

p<- net_score_plot(
nmt1, sucra1,
lab=trt1$description,
scl=1.1,ofs =0,
font.size = 0.3, pos=0.5,
labels="")

}
